# Speed up `bundler/setup` by using the raw `Gemfile.lock` information without extra processing whenever possible