## What was the end-user or developer problem that led to this PR?

We should make `bundler/setup` as fast as possible, since it's the cost that all scripts using Bundler to select versions of dependencies pay.

## What is your fix for the problem, implemented in this PR?

I noticed that even in the case where the `Gemfile` has no changes over the `Gemfile.lock` file, we're still doing some processing over the raw `Gemfile.lock` recorded resolution that did not seem necessary.

Further investigation revealed that this is only done for two edge cases:

* When dependencies are deleted from the Gemfile. In this case, we don't need to re-resolve, only to select the proper subset of the existing lockfile resolution.
* When the lockfile has redundant platform specific gems recorded. This can only happen to lockfiles generated with versions of Bundler prior to 1.16.1, due to [this bug](https://github.com/rubygems/bundler/issues/6491).

So my fix is to explicitly detect the above two situations and only do the extra processing there. Otherwise, use the raw lockfile.

This speeds bundling the Gemfile in https://github.com/technicalpickles/big-gemfile by about 5%, but should speed up all invocations of `bundler/setup` that don't need a re-resolve, the more dependencies, the more speed up.

```
➜  big-gemfile git:(main) ✗ hyperfine 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.3.17 ruby -rbundler/setup -e1'  
Benchmark 1: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     146.3 ms ±   0.9 ms    [User: 121.6 ms, System: 23.3 ms]
  Range (min … max):   144.5 ms … 148.4 ms    19 runs
 
Benchmark 2: BUNDLER_VERSION=2.3.17 ruby -rbundler/setup -e1
  Time (mean ± σ):     153.2 ms ±   0.7 ms    [User: 128.9 ms, System: 22.9 ms]
  Range (min … max):   151.5 ms … 154.3 ms    19 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1' ran
    1.05 ± 0.01 times faster than 'BUNDLER_VERSION=2.3.17 ruby -rbundler/setup -e1'
```

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x] Write code to solve the problem
- [x] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)